+++ apps/cowswap-frontend/src/modules/usdAmount/updaters/UsdPricesUpdater.ts
@@ -97,7 +97,7 @@ async function processQueue(queue: Token[], getUsdcPrice: () => Promise<Fraction
           state.updatedAt = Date.now()
         }
       } catch (e) {
-        console.debug(`[UsdPricesUpdater]: Failed to fetch price for `, currency.address)
+        console.debug(`[UsdPricesUpdater]: Failed to fetch price for`, currency.symbol)
       }
 
       return { [currency.address.toLowerCase()]: state }