+++ libs/wallet/src/web3-react/hooks/useWalletMetadata.ts
@@ -3,6 +3,7 @@ import { useMemo } from 'react'
 import { useWeb3React } from '@web3-react/core'
 
 import { useSafeAppsSdk } from './useSafeAppsSdk'
+import { useSafeAppsSdkInfo } from './useSafeAppsSdkInfo'
 
 import { useGnosisSafeInfo, useSelectedEip6963ProviderInfo } from '../../api/hooks'
 import { ConnectionType } from '../../api/types'
@@ -124,7 +125,10 @@ export function useIsSafeApp(): boolean {
  * regardless of the connection method (WalletConnect or inside Safe as an App)
  */
 export function useIsSafeWallet(): boolean {
-  return !!useGnosisSafeInfo()
+  const { safeAddress } = useSafeAppsSdkInfo() || {}
+  const hasSafeInfo = !!useGnosisSafeInfo()
+
+  return !!safeAddress || hasSafeInfo
 }
 
 /**